+++ R/docx_add.R
@@ -106,6 +106,8 @@ body_add_docx <- function( x, src, pos = "after" ){
 #' @param height height in inches
 #' @param res resolution of the png image in ppi
 #' @param scale Multiplicative scaling factor, same as in ggsave
+#' @param pos where to add the new element relative to the cursor,
+#' one of "after", "before", "on".
 #' @param ... Arguments to be passed to png function.
 #' @importFrom grDevices png dev.off
 #' @examples
@@ -121,7 +123,7 @@ body_add_docx <- function( x, src, pos = "after" ){
 #'   print(doc, target = tempfile(fileext = ".docx") )
 #' }
 #' @family functions for adding content
-body_add_gg <- function( x, value, width = 6, height = 5, res = 300, style = "Normal", scale = 1, ... ){
+body_add_gg <- function( x, value, width = 6, height = 5, res = 300, style = "Normal", scale = 1, pos = "after", ... ){
 
   if( !requireNamespace("ggplot2") )
     stop("package ggplot2 is required to use this function")
@@ -132,7 +134,7 @@ body_add_gg <- function( x, value, width = 6, height = 5, res = 300, style = "No
   print(value)
   dev.off()
   on.exit(unlink(file))
-  body_add_img(x, src = file, style = style, width = width, height = height)
+  body_add_img(x, src = file, style = style, width = width, height = height, pos = pos)
 }
 
 
@@ -248,13 +250,6 @@ body_add_par <- function( x, value, style = NULL, pos = "after" ){
 #' @family functions for adding content
 body_add_fpar <- function( x, value, style = NULL, pos = "after" ){
 
-  img_src <- sapply(value$chunks, function(x){
-    if( inherits(x, "external_img"))
-      as.character(x)
-    else NA_character_
-  })
-  img_src <- unique(img_src[!is.na(img_src)])
-
   if( !is.null(style) ){
     style_id <- get_style_id(data = x$styles, style=style, type = "paragraph")
   } else style_id <- NULL
@@ -350,6 +345,8 @@ body_add_toc <- function( x, level = 3, pos = "after", style = NULL, separator =
 #' @param width height in inches
 #' @param height height in inches
 #' @param res resolution of the png image in ppi
+#' @param pos where to add the new element relative to the cursor,
+#' one of "after", "before", "on".
 #' @param ... Arguments to be passed to png function.
 #' @importFrom grDevices png dev.off
 #' @examples
@@ -363,7 +360,7 @@ body_add_toc <- function( x, level = 3, pos = "after", style = NULL, separator =
 #'
 #' print(doc, target = tempfile(fileext = ".docx") )
 #' @family functions for adding content
-body_add_plot <- function( x, value, width = 6, height = 5, res = 300, style = "Normal", ... ){
+body_add_plot <- function( x, value, width = 6, height = 5, res = 300, style = "Normal", pos = "after", ... ){
 
   file <- tempfile(fileext = ".png")
   png(filename = file, width = width, height = height, units = "in", res = res, ...)
@@ -374,7 +371,7 @@ body_add_plot <- function( x, value, width = 6, height = 5, res = 300, style = "
     dev.off()
   } )
   on.exit(unlink(file))
-  body_add_img(x, src = file, style = style, width = width, height = height)
+  body_add_img(x, src = file, style = style, width = width, height = height, pos = pos)
 }
 
 
@@ -418,32 +415,26 @@ body_add_caption <- function( x, value, pos = "after"){
 #' @param pos where to add the new element relative to the cursor,
 #' one of "after", "before", "on".
 #' @keywords internal
-body_add_xml <- function(x, str, pos){
+body_add_xml <- function(x, str, pos = c("after", "before", "on")){
   xml_elt <- as_xml_document(str)
-  pos_list <- c("after", "before", "on")
+  pos <- match.arg(pos)
 
-  if( !pos %in% pos_list )
-    stop("unknown pos ", shQuote(pos, type = "sh"), ", it should be ",
-         paste( shQuote(pos_list, type = "sh"), collapse = " or ") )
-
-  cursor_elt <- x$doc_obj$get_at_cursor()
-
-  if( length(x) == 1 ){
-    xml_add_sibling(cursor_elt, xml_elt, .where = "before")
-    x <- cursor_end(x)
+  cursor_elt <- docx_current_block_xml(x)
+  if (is.null(cursor_elt)) {
+    xml_add_child(xml_find_first(x$doc_obj$get(), "/w:document/w:body"),
+                  xml_elt, .where = 0)
+    .name <- xml_name(xml_elt)
+    x$officer_cursor <- cursor_append(x$officer_cursor, .name)
+  } else if (pos == "on") {
+    xml_replace(cursor_elt, xml_elt)
+    x$officer_cursor <- cursor_replace_nodename(x$officer_cursor, xml_name(xml_elt))
+  } else if (pos == "after") {
+    xml_add_sibling(cursor_elt, xml_elt, .where = pos)
+    x$officer_cursor <- cursor_add_after(x$officer_cursor, xml_name(xml_elt))
   } else {
-    if( pos != "on")
-      xml_add_sibling(cursor_elt, xml_elt, .where = pos)
-    else {
-      xml_replace(cursor_elt, xml_elt)
-    }
-    if(pos == "after")
-      x <- cursor_forward(x)
-    else if(pos == "before" && !x$doc_obj$cursor_at_start())
-      x <- cursor_backward(x)
+    xml_add_sibling(cursor_elt, xml_elt, .where = pos)
+    x$officer_cursor <- cursor_add_before(x$officer_cursor, xml_name(xml_elt))
   }
-
-
   x
 }
 
@@ -469,7 +460,7 @@ body_add_xml2 <- function(x, str){
 #' doc <- body_add_par(doc, "centered text", style = "centered")
 #' doc <- body_bookmark(doc, "text_to_replace")
 body_bookmark <- function(x, id){
-  cursor_elt <- x$doc_obj$get_at_cursor()
+  cursor_elt <- docx_current_block_xml(x)
   ns_ <- "xmlns:w=\"http://schemas.openxmlformats.org/wordprocessingml/2006/main\""
   new_id <- UUIDgenerate()
   id <- check_bookmark_id(id)
@@ -517,17 +508,14 @@ body_bookmark <- function(x, id){
 #' print(my_doc, target = tempfile(fileext = ".docx"))
 body_remove <- function(x){
 
-  cursor_elt <- x$doc_obj$get_at_cursor()
+  cursor_elt <- docx_current_block_xml(x)
 
-  if( x$doc_obj$length() == 1 && xml_name(cursor_elt) == "sectPr"){
+  if(is.null(cursor_elt)){
     warning("There is nothing left to remove in the document")
     return(x)
   }
-
-  x$doc_obj$cursor_forward()
-  new_cursor_elt <- x$doc_obj$get_at_cursor()
+  x$officer_cursor <- cursor_delete(x$officer_cursor, xml_name(cursor_elt))
   xml_remove(cursor_elt)
-  x$doc_obj$set_cursor(xml_path(new_cursor_elt))
   x
 }