+++ R/docx_cursor.R
@@ -53,7 +53,11 @@
 #'   target = tempfile(fileext = ".docx")
 #' )
 cursor_begin <- function(x) {
-  x$doc_obj$cursor_begin()
+  if (length(x$officer_cursor$nodes_names) > 0L) {
+    x$officer_cursor$which <- 1L
+  } else {
+    x$officer_cursor$which <- 0L
+  }
   x
 }
 
@@ -76,7 +80,41 @@ cursor_begin <- function(x) {
 #'
 #' print(doc, target = tempfile(fileext = ".docx"))
 cursor_bookmark <- function(x, id) {
-  x$doc_obj$cursor_bookmark(id)
+  xpath_ <- sprintf("//w:bookmarkStart[@w:name='%s']", id)
+  bm_start <- xml_find_first(x$doc_obj$get(), xpath_)
+
+  if (inherits(bm_start, "xml_missing")) {
+    stop("cannot find bookmark ", shQuote(id), call. = FALSE)
+  }
+
+  bm_id <- xml_attr(bm_start, "id")
+
+  nodes_with_text <- xml_find_all(
+    x$doc_obj$get(),
+    "/w:document/w:body/*|/w:ftr/*|/w:hdr/*"
+  )
+  test_start <- sapply(nodes_with_text, function(node) {
+    expr <- sprintf("/descendant::w:bookmarkStart[@w:id='%s']", bm_id)
+    match_node <- xml_child(node, expr)
+    !inherits(match_node, "xml_missing")
+  })
+  if (!any(test_start)) {
+    stop("bookmark ", shQuote(id), " has not been found in the document", call. = FALSE)
+  }
+
+  test_end <- sapply(nodes_with_text, function(node) {
+    expr <- sprintf("/descendant::w:bookmarkEnd[@w:id='%s']", bm_id)
+    match_node <- xml_child(node, expr)
+    !inherits(match_node, "xml_missing")
+  })
+
+  on_same_par <- test_start == test_end
+  if (!all(on_same_par)) {
+    stop("bookmark ", shQuote(id), " does not end in the same paragraph (or is on the whole paragraph)", call. = FALSE)
+  }
+
+  x$officer_cursor$which <- which(test_start)[1]
+
   x
 }
 
@@ -86,7 +124,7 @@ cursor_bookmark <- function(x, id) {
 #' Set the cursor at the end of the document, on the last element
 #' of the document.
 cursor_end <- function(x) {
-  x$doc_obj$cursor_end()
+  x$officer_cursor$which <- length(x$officer_cursor$nodes_names)
   x
 }
 
@@ -98,7 +136,21 @@ cursor_end <- function(x) {
 #' that contains text specified in argument \code{keyword}.
 #' The argument \code{keyword} is a regexpr pattern.
 cursor_reach <- function(x, keyword) {
-  x$doc_obj$cursor_reach(keyword = keyword)
+  nodes_with_text <- xml_find_all(
+    x$doc_obj$get(),
+    "/w:document/w:body/*|/w:ftr/*|/w:hdr/*"
+  )
+
+  if (length(nodes_with_text) < 1) {
+    stop("no text found in the document", call. = FALSE)
+  }
+
+  text_ <- xml_text(nodes_with_text)
+  test_ <- grepl(pattern = keyword, x = text_)
+  if (!any(test_)) {
+    stop(keyword, " has not been found in the document", call. = FALSE)
+  }
+  x$officer_cursor$which <- which(test_)[1]
   x
 }
 
@@ -110,7 +162,18 @@ cursor_reach <- function(x, keyword) {
 #' that contains text specified in argument \code{keyword}.
 #' The argument \code{keyword} is a regexpr pattern.
 cursor_reach_test <- function(x, keyword) {
-  x$doc_obj$cursor_reachable(keyword = keyword)
+  nodes_with_text <- xml_find_all(
+    x$doc_obj$get(),
+    "/w:document/w:body/*|/w:ftr/*|/w:hdr/*"
+  )
+
+  if (length(nodes_with_text) < 1) {
+    stop("no text found in the document", call. = FALSE)
+  }
+
+  text_ <- xml_text(nodes_with_text)
+  test_ <- grepl(pattern = keyword, x = text_)
+  any(test_)
 }
 
 #' @export
@@ -118,7 +181,7 @@ cursor_reach_test <- function(x, keyword) {
 #' @section cursor_forward:
 #' Move the cursor forward, it increments the cursor in the document.
 cursor_forward <- function(x) {
-  x$doc_obj$cursor_forward()
+  x$officer_cursor$which <- min(c(length(x$officer_cursor$nodes_names), x$officer_cursor$which + 1L))
   x
 }
 
@@ -127,6 +190,73 @@ cursor_forward <- function(x) {
 #' @section cursor_backward:
 #' Move the cursor backward, it decrements the cursor in the document.
 cursor_backward <- function(x) {
-  x$doc_obj$cursor_backward()
+  x$officer_cursor$which <- max(c(0L, x$officer_cursor$which - 1L))
+  x
+}
+
+# officer docx cursor ----
+officer_cursor <- function(node) {
+  nodes <- xml_find_all(node, "/w:document/w:body/*")
+  nodes_names <- xml_name(nodes)
+  nodes_names <- nodes_names[!nodes_names %in% "sectPr"]
+  x <- list(
+    nodes_names = nodes_names,
+    which = length(nodes_names)
+  )
+  class(x) <- "officer_cursor"
+  x
+}
+as.character.officer_cursor <- function(x, ...) {
+  if (length(x$nodes_names) < 1) {
+    return(NA_character_)
+  }
+  paste0("/w:document/w:body/*[", x$which, "]")
+}
+
+## xml on cursor ----
+ooxml_on_cursor <- function(x, node) {
+  if (length(x$nodes_names) < 1) {
+    return(NULL)
+  }
+  node <- xml_find_first(node, as.character(x))
+  if (inherits(node, "xml_missing")) {
+    stop("cursor does not correspond to any node", call. = FALSE)
+  }
+  node
+}
+
+
+## cursor feed ----
+cursor_append <- function(x, what) {
+  x$nodes_names <- c(x$nodes_names, what)
+  x$which <- x$which + 1L
+  x
+}
+cursor_add_after <- function(x, what) {
+  seq_left <- seq_along(x$which)
+  set_left <- x$nodes_names[seq_left]
+  set_right <- x$nodes_names[-seq_left]
+  x$nodes_names <- c(set_left, what, set_right)
+  x$which <- x$which + 1L
+  x
+}
+
+cursor_add_before <- function(x, what) {
+  seq_left <- seq_along(x$which - 1)
+  set_left <- x$nodes_names[seq_left]
+  set_right <- x$nodes_names[-seq_left]
+  x$nodes_names <- c(set_left, what, set_right)
+  x
+}
+cursor_replace_nodename <- function(x, what) {
+  x$nodes_names[x$which] <- what
+  x
+}
+cursor_delete <- function(x, what) {
+  stopifnot(x$nodes_names[x$which] == what)
+  x$nodes_names <- x$nodes_names[-x$which]
+  if (x$which > length(x$nodes_names)) {
+    x$which <- x$which - 1L
+  }
   x
 }